% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{GetData}
\usage{
getData(url = "http://sandbox.ereuse.org/api/devices/",token = "f012ee230b324f384f4def745ed0a8702adc9452")
}
\arguments{
\item{url}{an uri pointing to a resource}

\item{token}{a string with a valid access for a given user and resource}
}
\value{
a list (result, error, url called)
}
\description{
\code{getData} This function is used to return data pointing to an url resource
}
\details{
These command will return any authorized data for a given uri resource.
}

