% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getDevices.R
\name{getDevicesJson}
\alias{getDevicesJson}
\title{Get Devices}
\usage{
getDevices(url)
}
\arguments{
\item{url}{an uri pointing to a RESTful API to provide a list of devices}
}
\value{
A list of 3 values. result is a json, error is a logical, call is list of one value (url called)
}
\description{
\code{getDevicesJson} This function returns all devices stored on GRD.
}
\details{
Not finished
}
\seealso{
http://sandbox.ereuse.org/swagger-ui/#!/default/get_api_devices
}

