% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getPath.R
\name{getPath}
\alias{getPath}
\title{GetPath}
\usage{
getPath(path_base = "http://sandbox.ereuse.org", name = "devices")
}
\arguments{
\item{path_base}{an uri pointing to a sandbox}

\item{name}{(string) of the API REST resource, for example: "agents", "devices", "events",...}
}
\value{
a list (result, error)
}
\description{
\code{getPath} This function is used to return the path to access API rest resources.
}
\details{
If name do not match to any resource returns an error.
}

