\name{Rgrd-package}
\alias{Rgrd-package}
\alias{Rgrd}
\docType{package}
\title{
Tools and Utilities for Grd
}
\description{
Data analysis and tools over a Global Record Device (GRD) system storing data about digital devices traceability.
}
\details{
\tabular{ll}{
Package: \tab Rgrd\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-08\cr
}
}
\author{
David Franquesa Griso

Maintainer: David Franquesa Griso <d.franquesa@ereuse.org>
}
\references{
~~ Literature or other references for background information ~~
}
\keyword{ package }
\seealso{
}
\examples{
~~ Some Examples ~~
}
