% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getData.R
\name{getData}
\alias{getData}
\title{GetData}
\usage{
getData(path.devices,token)
}
\arguments{
\item{url}{an uri pointing to a resource}

\item{token}{a string with a valid access for a given user and resource}
}
\value{
a list (result, error, url called)
}
\description{
\code{getData} This function is used to return data from an API REST where url is pointing
}
\details{
If name do not match to any resource returns an error.
}

