% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getToken.R
\name{getToken}
\alias{getToken}
\title{Get GRD Token}
\usage{
getToken(url, username, password)
}
\arguments{
\item{url}{an uri poiting to api-token, http://sandbox.ereuse.org/api-token-auth/}

\item{username}{an string with the username of a user}

\item{password}{an string}
}
\value{
a string qith token id
}
\description{
\code{getToken} uses getToken to get a token for GRD
}
\examples{
urlToken  <- "http://sandbox.ereuse.org/api-token-auth/"
 token     <- getToken(urlToken, username = "xxxx", password = "xxxx")
}

