% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getToken.R
\name{getToken}
\alias{getToken}
\title{Get Token}
\usage{
getToken(url = "xxx", username = "xxx", password = "xxx")
}
\arguments{
\item{url}{an uri pointing to a RESTful API to provide a token access}

\item{username}{(string) provided by ereuse.org federated platforms.}

\item{password}{(string) provided by ereuse.org federated platforms.}
}
\value{
a list (result, error, call())

result contains token id#'
}
\description{
\code{getToken} This function is used to get authentication token to access to your grd session. This will enable many bits of functionality as well as allow other commands to provide more options
}
\details{
If username or password is invalid HTTP Status Code = 400
}
\seealso{
http://sandbox.ereuse.org/swagger-ui/#!/default/post_api_token_auth
}

