\name{Rgrd-package}
\alias{Rgrd-package}
\alias{Rgrd}
\docType{package}
\title{
Data analysis over a Global Record Device system for electronic traceability
}
\description{
Some about
Rgrd Pakace
}
\details{
\tabular{ll}{
Package: \tab Rgrd\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2015-06-08\cr
}
urlToken  <- "http://sandbox.ereuse.org/api-token-auth/"
token     <- getToken(urlToken, username = "XXX", password = "XXX")
}
\author{
David Franquesa

Maintainer: David Franquesa <d.franquesa@ereuse.org>
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
}
\examples{
urlToken  <- "http://sandbox.ereuse.org/api-token-auth/"
token     <- getToken(urlToken, username = "XXX", password = "XXX")
}
